update product endpoint

+++ src/controllers/sellerController.js
@@ -125,7 +125,7 @@ const addNewProduct = async (req, res) => {
 const updateProduct = async (req, res) => {
     try {
         const updateFields = {};
-        const { productId, name, category, brand, productDescription, variants } = req.body;
+        const { productId, name, category, brand, productDescription, variants, productImages } = req.body;
 
         const product = await ProductService.getProductById(productId);
 
@@ -134,25 +134,20 @@ const updateProduct = async (req, res) => {
         if (brand) updateFields.brand = brand;
         if (productDescription) updateFields.productDescription = productDescription;
         if (variants) updateFields.variants = JSON.parse(variants);
+        if (productImages) {
+            const parsedProductImages = JSON.parse(productImages);
 
-        if (req.files) {
-            if (req.files['thumbnailImage']) {
-                if (product.thumbnailImage) {
-                    deleteFile(product.thumbnailImage);
-                }
-                updateFields.thumbnailImage = req.files['thumbnailImage'][0].path;
+            const oldImages = product.productImages;
+            const imagesToDelete = oldImages.filter((img) => !parsedProductImages.includes(img));
+            imagesToDelete.forEach((img) => deleteFile(img));
+            updateFields.productImages = parsedProductImages;
+        }
+
+        if (req.files && req.files['thumbnailImage']) {
+            if (product.thumbnailImage && req.files['thumbnailImage'][0].path !== product.thumbnailImage) {
+                deleteFile(product.thumbnailImage);
             }
-
-            const productImages = Object.keys(req.files)
-                .filter((fieldName) => fieldName !== 'thumbnailImage')
-                .flatMap((fieldName) => req.files[fieldName].map((file) => file.path));
-
-            if (productImages.length > 0) {
-                const oldImages = product.productImages;
-                const imagesToDelete = oldImages.filter((img) => !productImages.includes(img));
-                imagesToDelete.forEach((img) => deleteFile(img));
-                updateFields.productImages = productImages;
-            }
+            updateFields.thumbnailImage = req.files['thumbnailImage'][0].path;
         }
 
         const updatedProduct = await ProductService.updateProduct(productId, updateFields);
